refactor arcana db
