+++ components/Arcana/styles.module.css
@@ -1,0 +1,4 @@
+.link a {
+	text-decoration: underline;
+	@apply text-emerald-400;
+}
