+++ pages/arcana.mdx
@@ -27,191 +27,114 @@ import { DbArcana } from "@/db";
 
 ### The Sorceress [#sorceress]
 
-<Arcana data={DbArcana.sorceress}>
-  While you **Channel** your **Ω Moves**, everything moves slower for **4 Sec.**
-</Arcana>
+<Arcana data={DbArcana.sorceress} />
 
 ### The Wayward Son [#wayward-son]
 
-<Arcana data={DbArcana.waywardSon}>
-  After you exit a **Location**, restore **4** <IconLife />, or twice that if
-  you have no more than **30%** <IconLife />.
-</Arcana>
+<Arcana data={DbArcana.waywardSon} />
 
 ### The Huntress [#huntress]
 
 <Arcana data={DbArcana.huntress}>
-  While you have less than **100%** <IconMagick />, your **Attack** and
-  **Special** deal **+50%** damage.
+
 </Arcana>
 
 ### Night [#night]
 
-<Arcana data={DbArcana.night}>
-  Your **Ω Cast** has **+100 Power** whenever it detonates.
-
-</Arcana>
+<Arcana data={DbArcana.night} />
 
 ### The Moon [#moon]
 
-<Arcana data={DbArcana.moon}>
-Your **Hex** also charges up automatically as though you used **4** <IconMagick /> every **1 Sec.**
-
-**Awakening**: _Activate **any** surrounding Card. ([Night](/arcana#night), [The Unseen](/arcana#unseen), [Death](/arcana#death))_
-
-</Arcana>
+<Arcana data={DbArcana.moon} />
 
 ## Row 2
 
 ### The Furies [#furies]
 
-<Arcana data={DbArcana.furies}>
-  Deal **+30%** damage to foes in your **Casts**.
-</Arcana>
+<Arcana data={DbArcana.furies} />
 
 ### The Titan [#titan]
 
-<Arcana data={DbArcana.titan}>
-  Gain **+40** <IconMaxLife /> and **+40** <IconMaxMagick />.
-</Arcana>
+<Arcana data={DbArcana.titan} />
 
 ### The Messenger [#messenger]
 
-<Arcana data={DbArcana.messenger}>
-You have **+5%** chance to **Dodge**.
-
-</Arcana>
+<Arcana data={DbArcana.messenger} />
 
 ### The Unseen [#unseen]
 
-<Arcana data={DbArcana.unseen}>
-  Restore **6** <IconMagick /> every **1 Sec.**
-</Arcana>
+<Arcana data={DbArcana.unseen} />
 
 ### Death [#death]
 
-<Arcana data={DbArcana.death}>
-  Your **Ω Moves** have **+15%** chance to deal **Critical** damage if your last
-  one was different.
-</Arcana>
+<Arcana data={DbArcana.death} />
 
 ## Row 3
 
 ### The Swift Runner [#swift-runner]
 
-<Arcana data={DbArcana.swiftRunner}>
-  Your **Dash** is immediate and your **Sprint** is **+10%** faster.
-</Arcana>
+<Arcana data={DbArcana.swiftRunner} />
 
 ### Eternity [#eternity]
 
-<Arcana data={DbArcana.eternity}>
-  Set forth each night with **3** <IconDeathDefiance />.
-</Arcana>
+<Arcana data={DbArcana.eternity} />
 
 Don't leave home without it. As close to a must-pick as you can get.
 
 ### The Centaur [#centaur]
 
-<Arcana data={DbArcana.centaur}>
-After every **5 Locations**, gain **+5** <IconMaxLife /> and **+5** <IconMaxMagick />.
-
-**Awakening**: _Activate Cards that use **1** <IconGrasp /> through **5** <IconGrasp />._
-
-</Arcana>
+<Arcana data={DbArcana.centaur} />
 
 ### Origination [#origination]
 
-<Arcana data={DbArcana.origination}>
-  Deal **+50%** damage to foes afflicted with at least **2 Curse** effects from
-  different Olympians.
-</Arcana>
+<Arcana data={DbArcana.origination} />
 
 ### The Lovers [#lovers]
 
-<Arcana data={DbArcana.lovers}>
-  In **Guardian Encounters**, you take **0** damage the first **3** time(s) you
-  are hit.
-</Arcana>
+<Arcana data={DbArcana.lovers} />
 
 ## Row 4
 
 ### The Fates [#fates]
 
-<Arcana data={DbArcana.fates}>
-  Set forth each night with **+3** <IconReroll />. You can alter **Location
-  Rewards**.
-</Arcana>
+<Arcana data={DbArcana.fates} />
 
 ### The Boatman [#boatman]
 
-<Arcana data={DbArcana.boatman}>
-  Set forth each night with **+300** <IconGold />.
-</Arcana>
+<Arcana data={DbArcana.boatman} />
 
 ### The Artificer [#artificer]
 
 <Arcana data={DbArcana.artificer}>
-  Set forth each night with **3** chance(s) to turn any **Minor Find** into a
-  random **Major Find**.
+
 </Arcana>
 
 ### Excellence [#excellence]
 
-<Arcana data={DbArcana.excellence}>
-  Any **Boons** you are offered have **+50%** chance to be improved to
-  <TextHighlight type="rare">**Rare**</TextHighlight>.
-</Arcana>
+<Arcana data={DbArcana.excellence} />
 
 ### The Queen [#queen]
 
-<Arcana data={DbArcana.queen}>
-Any **Boons** you are offered have **+10%** chance to be a <TextHighlight type='duo'>**Duo**</TextHighlight> _(whenever possible)_.
-
-**Awakening**: _Activate no more than **2** Cards that use the same amount of **Grasp**._
-
-</Arcana>
+<Arcana data={DbArcana.queen} />
 
 ## Row 5
 
 ### The Seer [#seer]
 
-<Arcana data={DbArcana.seer}>
-  Set forth each night with +4 <IconReroll />.
-
-**Awakening**: _Activate **all** surrounding Cards. ([The Fates](/arcana#the-fates),
-[The Boatman](/arcana#the-boatman), [The Champions](/arcana#the-champions))_
-
-</Arcana>
+<Arcana data={DbArcana.seer} />
 
 ### The Champions [#champions]
 
-<Arcana data={DbArcana.champions}>
-  Set forth each night with **+3** <IconReroll />. You can alter **Boons** and
-  certain other choices.
-</Arcana>
+<Arcana data={DbArcana.champions} />
 
 ### Strength [#strength]
 
-<Arcana data={DbArcana.strength}>
-  While you have no more than **30%** <IconLife />, you take **-50%** damage and
-  deal **+50%**.
-</Arcana>
+<Arcana data={DbArcana.strength} />
 
 ### Divinity [#divinity]
 
-<Arcana data={DbArcana.divinity}>
-  Any **Boons** you are offered have **+20%** chance to be improved to <TextHighlight type='epic'>**Epic**</TextHighlight>.
-
-**Awakening**: _Activate all **5** Cards in any other row or column._
-
-</Arcana>
+<Arcana data={DbArcana.divinity} />
 
 ### Judgment [#judgment]
 
-<Arcana data={DbArcana.judgment}>
-After you vanquish a **Guardian**, activate **5** random inactive **Arcana Cards**.
-
-**Awakening**: _Activate no more than **3** Cards._
-
-</Arcana>
+<Arcana data={DbArcana.judgment} />
